Update Cassiere, SupermercatoImpl e Scaffalista

+++ simulazionesupermercato/src/main/melle/lavoratori/Cassiere.java
@@ -4,7 +4,6 @@ import java.util.List;
 
 import main.melle.clienti.Cliente;
 import main.melle.supermercato.Supermercato;
-import main.melle.supermercato.SupermercatoImpl;
 import main.melle.zonacasse.Cassa;
 import main.miftari.prodotti.Prodotto;
 
@@ -12,7 +11,7 @@ public class Cassiere extends Lavoratore{
 
 	private Cassa cassaCheGestisco;
 	
-	public Cassiere(int stipendio, Cassa cassaCheGestisco) {
+	public Cassiere(double stipendio, Cassa cassaCheGestisco) {
 		super(stipendio);
 		this.cassaCheGestisco = cassaCheGestisco;
 	}
@@ -36,7 +35,6 @@ public class Cassiere extends Lavoratore{
 					List<Prodotto> prodottiDaPagare = daServire.getProdottiPresi();
 					double prezzoTotale = this.cassaCheGestisco.getRegistratoreDiCassa().calcolaCostoSpesa(null);
 					this.cassaCheGestisco.getRegistratoreDiCassa().registraPagamento(prodottiDaPagare, prezzoTotale);
-					// che ne è dei prodotti pagati??
 					// adesso si può togliere il cliente dalla coda
 					this.cassaCheGestisco.togliDallaCoda(daServire);
 					// quando nessuno ha più il riferimento a cliente, questo oggetto verrà eliminato dal garbage collector
